Cache extractor results while offline (fixes #31)
Instead of using temporary files, use a JSON file with a filename
of the md5 of the audio file's filename. If unable to reach the
acousticbrainz server for whatever reason, save the result and mark
it as "offline" in the database. When processing that file again
on a subsequent run, check if it was marked as "offline" and attempt
to re-submit it if possible.